# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Resource Booking System API",
        "description": "API documentation for the Resource Booking Management System",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/bookings": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "List all bookings",
                "operationId": "4535e967ef15bc76dadf1efa40192f0d",
                "parameters": [
                    {
                        "name": "resource_id",
                        "in": "query",
                        "description": "Filter bookings by resource ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Filter bookings by user ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter bookings by status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "confirmed",
                                "cancelled"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of bookings",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Booking"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Create a new booking",
                "operationId": "43e6dcca842886074084b8db86af7802",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreBookingRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Booking created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Booking"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "409": {
                        "description": "Booking time slot is not available"
                    }
                }
            }
        },
        "/api/bookings/{id}": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Get a specific booking",
                "operationId": "c383a28985575aa1a16f783b2f762554",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Booking"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Booking not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Update a specific booking",
                "operationId": "de1ebc78c3adcb535d12d1c8dae98055",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateBookingRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Booking updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Booking"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Cancel a booking",
                "operationId": "a53474f63d3a2fe2275fc0b6c8cc16cb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Booking cancelled successfully"
                    },
                    "404": {
                        "description": "Booking not found"
                    },
                    "422": {
                        "description": "Cannot cancel past booking"
                    }
                }
            }
        },
        "/api/resources/{id}/bookings": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Get all bookings for a specific resource",
                "operationId": "bffd2f2a709c738d85c922d33be2f20a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of bookings",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Booking"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Booking": {
                "title": "Booking",
                "description": "Booking model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID of the booking",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "resource_id": {
                        "title": "Resource ID",
                        "description": "ID of the resource being booked",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "user_id": {
                        "title": "User ID",
                        "description": "ID of the user making the booking",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "start_time": {
                        "title": "Start Time",
                        "description": "Start time of the booking",
                        "type": "string",
                        "format": "datetime",
                        "example": "2025-01-01 10:00:00"
                    },
                    "end_time": {
                        "title": "End Time",
                        "description": "End time of the booking",
                        "type": "string",
                        "format": "datetime",
                        "example": "2025-01-01 11:00:00"
                    },
                    "status": {
                        "title": "Status",
                        "description": "Status of the booking",
                        "type": "string",
                        "enum": [
                            "confirmed",
                            "cancelled"
                        ],
                        "example": "confirmed"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2025-01-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2025-01-01T00:00:00.000000Z"
                    },
                    "resource": {
                        "$ref": "#/components/schemas/Resource"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Booking"
                }
            },
            "Resource": {
                "title": "Resource",
                "description": "Resource model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID of the resource",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "Name",
                        "description": "Name of the resource",
                        "type": "string",
                        "example": "Meeting Room A"
                    },
                    "type": {
                        "title": "Type",
                        "description": "Type of the resource",
                        "type": "string",
                        "example": "room"
                    },
                    "description": {
                        "title": "Description",
                        "description": "Description of the resource",
                        "type": "string",
                        "example": "A spacious meeting room with projector"
                    },
                    "is_active": {
                        "title": "Is Active",
                        "description": "Whether the resource is active",
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2025-01-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2025-01-01T00:00:00.000000Z"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Resource"
                }
            },
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID of the user",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "Name",
                        "description": "Name of the user",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "title": "Email",
                        "description": "Email of the user",
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "email_verified_at": {
                        "title": "Email Verified At",
                        "description": "Datetime when email was verified",
                        "type": "string",
                        "format": "datetime",
                        "example": "2025-01-01T00:00:00.000000Z"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2025-01-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2025-01-01T00:00:00.000000Z"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            },
            "StoreBookingRequest": {
                "title": "Store Booking Request",
                "description": "Store Booking request body data",
                "required": [
                    "resource_id",
                    "user_id",
                    "start_time",
                    "end_time"
                ],
                "properties": {
                    "resource_id": {
                        "title": "resource_id",
                        "description": "ID of the resource to book",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "user_id": {
                        "title": "user_id",
                        "description": "ID of the user making the booking",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "start_time": {
                        "title": "start_time",
                        "description": "Start time of the booking",
                        "type": "string",
                        "format": "datetime",
                        "example": "2025-01-01 10:00:00"
                    },
                    "end_time": {
                        "title": "end_time",
                        "description": "End time of the booking",
                        "type": "string",
                        "format": "datetime",
                        "example": "2025-01-01 11:00:00"
                    }
                },
                "type": "object"
            },
...
            "UpdateBookingRequest": {
                "title": "Update Booking Request",
                "description": "Update Booking request body data",
                "properties": {
                    "start_time": {
                        "title": "start_time",
                        "description": "Start time of the booking",
                        "type": "string",
                        "format": "datetime",
                        "example": "2025-01-01 10:00:00"
                    },
                    "end_time": {
                        "title": "end_time",
                        "description": "End time of the booking",
                        "type": "string",
                        "format": "datetime",
                        "example": "2025-01-01 11:00:00"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}